app/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.2.2
django-filter
setuptools
markdown
